feature/layout-refactoring move components from layout
