--- src/app/layouts/Snackbar/index.ts
@@ -1,3 +1,0 @@
-import {Snackbar} from './Snackbar';
-
-export default Snackbar;
