+++ src/app/snackbar/index.ts
@@ -1,0 +1,3 @@
+import {Snackbar} from './snackbar';
+
+export default Snackbar;
